[Refactoring] Move LocalizeString to its own file
