+++ lib/Refactoring/Refactoring.cpp
@@ -566,44 +566,6 @@ collectRefactoringsAtCursor(SourceFile *SF, unsigned Line, unsigned Column,
   return collectRefactorings(Tok, /*ExcludeRename=*/false);
 }
 
-static Expr *findLocalizeTarget(ResolvedCursorInfoPtr CursorInfo) {
-  auto ExprStartInfo = dyn_cast<ResolvedExprStartCursorInfo>(CursorInfo);
-  if (!ExprStartInfo)
-    return nullptr;
-  struct StringLiteralFinder: public SourceEntityWalker {
-    SourceLoc StartLoc;
-    Expr *Target;
-    StringLiteralFinder(SourceLoc StartLoc): StartLoc(StartLoc), Target(nullptr) {}
-    bool walkToExprPre(Expr *E) override {
-      if (E->getStartLoc() != StartLoc)
-        return false;
-      if (E->getKind() == ExprKind::InterpolatedStringLiteral)
-        return false;
-      if (E->getKind() == ExprKind::StringLiteral) {
-        Target = E;
-        return false;
-      }
-      return true;
-    }
-  } Walker(ExprStartInfo->getTrailingExpr()->getStartLoc());
-  Walker.walk(ExprStartInfo->getTrailingExpr());
-  return Walker.Target;
-}
-
-bool RefactoringActionLocalizeString::isApplicable(ResolvedCursorInfoPtr Tok,
-                                                   DiagnosticEngine &Diag) {
-  return findLocalizeTarget(Tok);
-}
-
-bool RefactoringActionLocalizeString::performChange() {
-  Expr* Target = findLocalizeTarget(CursorInfo);
-   if (!Target)
-    return true;
-  EditConsumer.accept(SM, Target->getStartLoc(), "NSLocalizedString(");
-  EditConsumer.insertAfter(SM, Target->getEndLoc(), ", comment: \"\")");
-  return false;
-}
-
 struct MemberwiseParameter {
   CharSourceRange NameRange;
   Type MemberType;
